Fix the colour set difference, a part of the algorithm and output.
The set difference resulted in the resulting set looking like
{ 0, x, y, ... } if cSet was {1, 0}. This is because cSet was not
sorted, so this commit ensures that it is. Furthermore, when looking
at neighbours, we only actually care about the ones we've put back
in, not all of them, so we keep a list of popped nodes instead of
looking into adjMatrix. In addition to that, make the output of
the clash graph a bit more legible.