Initial commit adding the testing infrastructure and a simple test.
The test that was added has a lot of code that will eventually
likely be abstracted away into a more general framework to avoid
a lot of copy paste.

The that was added is equivalent to "testfile" in example1, but instead
of just printing the result, it checks whether or not there is any case
after the allocation where two simultaneously live virtual registers are
allocated to the same physical register.