Implement a policy-based setup for tests.
Currently, there are some issues with the build, namely that meson
seems to order the linker flags wrong and they end up being ignored
by clang. This code, however, does build given a C++17-compatible
compiler.

This effectively means that the expected version of C++ is now bumped to
C++17.